requirements.txt
gmqtt
aiosqlite
pyyaml
python-dotenv
psutil